Create new google-cloud-resource_manager gem package
