+++ google-cloud-resource_manager/.yardopts
@@ -1,0 +1,9 @@
+--no-private
+--title=Google Cloud Resource Manager
+--markup markdown
+
+./lib/**/*.rb
+-
+README.md
+OVERVIEW.md
+AUTHENTICATION.md
