+++ google-cloud-resource_manager/lib/google/cloud/resource_manager/credentials.rb
@@ -23,9 +23,9 @@ module Google
       class Credentials < Google::Cloud::Credentials
         SCOPE = ["https://www.googleapis.com/auth/cloud-platform"]
         PATH_ENV_VARS = %w(RESOURCE_MANAGER_KEYFILE
-                           GCLOUD_KEYFILE GOOGLE_CLOUD_KEYFILE)
-        JSON_ENV_VARS = %w(RESOURCE_MANAGER_KEYFILE_JSON GCLOUD_KEYFILE_JSON
-                           GOOGLE_CLOUD_KEYFILE_JSON)
+                           GOOGLE_CLOUD_KEYFILE GCLOUD_KEYFILE)
+        JSON_ENV_VARS = %w(RESOURCE_MANAGER_KEYFILE_JSON
+                           GOOGLE_CLOUD_KEYFILE_JSON GCLOUD_KEYFILE_JSON)
       end
     end
   end
