+++ google-cloud-resource_manager/lib/google/cloud/resource_manager/manager.rb
